Kill warnings

+++ avs_core/convert/convert_rgb.cpp
@@ -433,9 +433,9 @@ PackedRGBtoPlanarRGB::PackedRGBtoPlanarRGB(PClip src, bool _sourceHasAlpha, bool
 template<typename pixel_t, bool targetHasAlpha>
 // minimum width: 32 bytes (8 RGBA pixels for 8 bits, 4 RGBA pixels for 16 bits)
 static void convert_rgba_to_rgbp_sse2(const BYTE *srcp, BYTE * (&dstp)[4], int src_pitch, int (&dst_pitch)[4], size_t width, size_t height) {
-  const int rowsize = width * sizeof(pixel_t) * 4;
+  const size_t rowsize = width * sizeof(pixel_t) * 4;
   const int pixels_at_a_time = (sizeof(pixel_t) == 1) ? 8 : 4;
-  const int wmod = (width / pixels_at_a_time) * pixels_at_a_time; // 8 pixels for 8 bit, 4 pixels for 16 bit
+  const size_t wmod = (width / pixels_at_a_time) * pixels_at_a_time; // 8 pixels for 8 bit, 4 pixels for 16 bit
   for (size_t y = height; y > 0; --y) {
     __m128i BGRA_lo, BGRA_hi;
     for (int x = 0; x < wmod; x+=pixels_at_a_time) {
@@ -466,7 +466,7 @@ static void convert_rgba_to_rgbp_sse2(const BYTE *srcp, BYTE * (&dstp)[4], int s
     if (wmod != width) {
       // width = 17: 0..7 8..15, 16
       // last_start = 1 (9..16 8 pixels)  width - pixels_at_a_time
-      int last_start = (width - pixels_at_a_time);
+      size_t last_start = (width - pixels_at_a_time);
       BGRA_lo = _mm_loadu_si128(reinterpret_cast<const __m128i *>(srcp + last_start * 32 / pixels_at_a_time));
       BGRA_hi = _mm_loadu_si128(reinterpret_cast<const __m128i *>(srcp + last_start * 32 / pixels_at_a_time + 16));
       __m128i pack_lo, pack_hi, eightbytes_of_pixels;
@@ -595,7 +595,7 @@ template<typename pixel_t, int target_numcomponents, bool hasSrcAlpha>
 static void convert_rgbp_to_rgb_sse2(const BYTE *(&srcp)[4], BYTE * dstp, int (&src_pitch)[4], int dst_pitch, size_t width, size_t height) {
   const int pixels_at_a_time = 8 / sizeof(pixel_t); // 8x uint8_t, 4xuint16_t 8 bytes
 
-  const int wmod = (width / pixels_at_a_time) * pixels_at_a_time;
+  const size_t wmod = (width / pixels_at_a_time) * pixels_at_a_time;
 
   const __m128i transparent = _mm_set1_epi8((char)0xFF); // 0xFFFF for uint16_t
 
@@ -636,7 +636,7 @@ static void convert_rgbp_to_rgb_sse2(const BYTE *(&srcp)[4], BYTE * dstp, int (&
       // last_start = 1 (9..16 8 pixels)  width - pixels_at_a_time
       // width = 8 -> x = 0
       //         9 -> x = 1
-      int x = (width - pixels_at_a_time);
+      size_t x = (width - pixels_at_a_time);
       __m128i R, G, B, A;
       G = _mm_loadl_epi64(reinterpret_cast<const __m128i * > (srcp[0] + x*sizeof(pixel_t))); // 8 bytes G7..G0 or G3..G0
       B = _mm_loadl_epi64(reinterpret_cast<const __m128i * > (srcp[1] + x*sizeof(pixel_t))); // 8 bytes
